# onefrontend/web/angular.frontend/src/app/shared/models/result.model.ts
export class StringResult {
  public message: string;
  public result: string;
}

export class Result<T> {
  public success: boolean;
  public message: string;
  public value: T;
}

export class ListResult<T> extends Result<T> {
  public count:  number;
}